requirements.txt
tkinter
openai
python-dotenv 